[onert] Unify Managing tensors of controlflow and cpu
This commit unifies managing tensors of controlflow and cpu.
