+++ runtime/onert/backend/cpu/TensorBuilder.cc
@@ -36,20 +36,22 @@ TensorBuilder::TensorBuilder()
 }
 
 void TensorBuilder::registerTensorInfo(const ir::OperandIndex &ind, const ir::OperandInfo &info,
-                                       ir::Layout, bool as_const)
+                                       ir::Layout layout, bool as_const)
 {
   _tensor_info_map.emplace(ind, info);
 
   if (as_const)
     _constants.append(ind);
 
+  // CPU backend supports only one layout as NHWC
+  assert(layout == ir::Layout::NHWC);
   if (info.isDynamic())
   {
-    _dynamic_tensor_mgr->buildTensor(ind, info);
+    _dynamic_tensor_mgr->buildTensor(ind, info, layout);
   }
   else
   {
-    _static_tensor_mgr->buildTensor(ind, info, _constants.contains(ind));
+    _static_tensor_mgr->buildTensor(ind, info, layout, _constants.contains(ind));
   }
 }
