+++ runtime/onert/core/src/backend/controlflow/TensorBuilder.h
@@ -17,10 +17,10 @@
 #ifndef __ONERT_BACKEND_CONTROLFLOW_TENSOR_BUILDER_H__
 #define __ONERT_BACKEND_CONTROLFLOW_TENSOR_BUILDER_H__
 
-#include "DynamicTensorManager.h"
-#include "StaticTensorManager.h"
-#include "TensorRegistry.h"
-#include "operand/Tensor.h"
+#include <backend/cpu_common/DynamicTensorManager.h>
+#include <backend/cpu_common/StaticTensorManager.h>
+#include <backend/cpu_common/TensorRegistry.h>
+#include <backend/cpu_common/Tensor.h>
 
 #include <backend/ITensorBuilder.h>
 #include <ir/OperandIndexMap.h>
@@ -80,14 +80,14 @@ public:
    *        If not, program will crash with assert or exception.
    * @return shared_ptr<operand::Tensor>
    */
-  std::shared_ptr<operand::Tensor> at(const ir::OperandIndex &ind);
+  std::shared_ptr<cpu_common::Tensor> at(const ir::OperandIndex &ind);
 
   std::shared_ptr<ITensorRegistry> tensorRegistry() override { return _tensor_reg; }
 
 private:
-  const std::shared_ptr<TensorRegistry> _tensor_reg;
-  std::unique_ptr<StaticTensorManager> _static_tensor_mgr;
-  std::unique_ptr<DynamicTensorManager> _dynamic_tensor_mgr;
+  const std::shared_ptr<cpu_common::TensorRegistry> _tensor_reg;
+  std::unique_ptr<cpu_common::StaticTensorManager> _static_tensor_mgr;
+  std::unique_ptr<cpu_common::DynamicTensorManager> _dynamic_tensor_mgr;
   ir::OperandIndexMap<ir::OperandInfo> _tensor_info_map;
   ir::OperandIndexMap<ir::Layout> _tensor_layout_map;
   ir::OperandIndexSequence _constants;
